Versao para teste com container

+++ agent/RunAgent.ps1
@@ -7,7 +7,9 @@ param(
 
 Write-Host "Starting: $($MyInvocation.MyCommand.Definition)"
 
-Remove-Item $Env:AZP_TOKEN
+if (-not [string]::IsNullOrWhiteSpace($env:AZP_TOKEN)) {
+  Remove-Item $Env:AZP_TOKEN
+}
 
 # $dotnetRoot=""
 # [Environment]::SetEnvironmentVariable("DOTNET_ROOT", $dotnetRoot, 'Machine')
@@ -16,4 +18,4 @@ Set-Location $pathAgent
     
 Write-Host "Executing run.cmd in $PWD ..."
 
-#./run.cmd
+./run.cmd
